requirements.txt
python-arango
mysql-connector
